pyproject.toml
[tool.poetry]
name = "shepardtone"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
httpx = "^0.27.0"
attrs = "^23.2.0"
cattrs = "^23.2.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
respx = "^0.21.1"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
